# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Financial OS API
  version: "1.0.0"
  description: |
    Private single-owner receipt-capture API.

    All `/api/v1/` routes require a Firebase-issued Bearer token for the
    allowlisted owner identity (IAM-01, IAM-02).

    All `/internal/v1/` routes accept only Cloud Tasks OIDC-authenticated
    requests with the correct audience (QUE-01). They must not be reachable
    from the public internet.

    Changes to this contract after Wave 1 require supervisor approval,
    impact analysis, and affected-agent notification.
  contact:
    name: Financial OS — supervisor
  x-wave: "1"
  x-baseline: "planning-baseline-2026-08-12-r1"
servers:
  - url: https://api.example.invalid
    description: >
      Placeholder. Real server URL is a non-public Cloud Run service URL
      managed through Secret Manager and never committed to this file.
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase JWT (Google identity, server-validated, owner allowlist enforced)
    CloudTasksOIDC:
      type: http
      scheme: bearer
      bearerFormat: >
        Google OIDC JWT issued by Cloud Tasks service account.
        Audience must match the worker service URL. Validated server-side.
  schemas:
...
    UUID:
      type: string
      format: uuid
      description: Version 4 UUID in lowercase canonical form (8-4-4-4-12).
      example: "a1b2c3d4-e5f6-4a7b-8c9d-e0f1a2b3c4d5"
    Timestamp:
      type: string
      format: date-time
      description: UTC timestamp in RFC 3339 format.
      example: "2026-08-12T14:30:00.000Z"
...
    ProcessingStatus:
      type: string
      enum:
        - reserved
        - uploading
        - uploaded
        - queued
        - processing
        - extracted
        - retryable_failed
        - failed
        - abandoned
      description: |
        Internal processing lifecycle. Public API may collapse early states
        into simpler language but never implies completion that has not occurred.

        State machine:
          reserved → uploading → uploaded → queued → processing → extracted
                                                                ↘ retryable_failed → queued
                                                                ↘ failed
          reserved | uploading → abandoned (reconciliation sweep only)
    VerificationStatus:
      type: string
      enum:
        - unreviewed
        - system_validated
        - needs_review
        - human_verified
      description: >
        Independent trust dimension. Processing and verification states advance
        independently. No automated process sets human_verified.
    FinancialContext:
      type: string
      enum:
        - personal
        - rental_property
      description: >
        Coarse financial context. Defaults to personal. rental_property is
        available for exclusion from personal behavior aggregates; full
        property modeling is deferred.
...
    AssetInput:
      type: object
      required:
        - ordinal
        - declared_mime_type
        - byte_size
      properties:
        ordinal:
          type: integer
          minimum: 1
          description: 1-based position of this image in the ordered receipt.
        declared_mime_type:
          type: string
          description: >
            Client-declared MIME type. Server treats this as a claim only.
            Verified server-side at finalization (OBJ-03).
          example: "image/jpeg"
        byte_size:
          type: integer
          minimum: 1
          maximum: 10485760
          description: Declared byte size. Server verifies at finalization.
    UploadCapability:
      type: object
      required:
        - asset_id
        - ordinal
        - upload_url
        - method
        - expires_at
        - allowed_mime_types
      properties:
        asset_id:
          $ref: "#/components/schemas/UUID"
        ordinal:
          type: integer
          minimum: 1
        upload_url:
          type: string
          format: uri
          description: >
            Short-lived server-selected signed URL for direct PUT upload to
            private object storage. This is a bearer secret — never log,
            cache, or emit in analytics (OBJ-02, LOG-01).
        method:
          type: string
          enum: ["PUT"]
        expires_at:
          $ref: "#/components/schemas/Timestamp"
        allowed_mime_types:
          type: array
          items:
            type: string
          description: MIME types accepted at this upload path.
...
    RevisionSummary:
      type: object
      properties:
        revision_id:
          $ref: "#/components/schemas/UUID"
        source_type:
          type: string
          enum: [extractor, human, import]
        merchant_normalized:
          type:
            - string
            - "null"
        purchase_datetime:
          type:
            - string
            - "null"
          format: date-time
        currency:
          $ref: "#/components/schemas/Currency"
        subtotal_minor:
          $ref: "#/components/schemas/MinorUnits"
        tax_minor:
          $ref: "#/components/schemas/MinorUnits"
        tip_minor:
          $ref: "#/components/schemas/MinorUnits"
        discount_minor:
          $ref: "#/components/schemas/MinorUnits"
        total_minor:
          $ref: "#/components/schemas/MinorUnits"
        overall_confidence:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1

    # ── Receipt schemas ───────────────────────────────────────────────────────
    ReceiptListItem:
      type: object
      required:
        - receipt_id
        - processing_status
        - verification_status
        - financial_context
        - expected_asset_count
        - created_at
      properties:
        receipt_id:
          $ref: "#/components/schemas/UUID"
        processing_status:
          $ref: "#/components/schemas/ProcessingStatus"
        verification_status:
          $ref: "#/components/schemas/VerificationStatus"
        financial_context:
          $ref: "#/components/schemas/FinancialContext"
        expected_asset_count:
          type: integer
          minimum: 1
        acknowledged_at:
          type:
            - string
            - "null"
          format: date-time
        created_at:
          $ref: "#/components/schemas/Timestamp"
        current_revision:
          oneOf:
            - $ref: "#/components/schemas/RevisionSummary"
            - type: "null"
    ReceiptDetail:
      allOf:
        - $ref: "#/components/schemas/ReceiptListItem"
        - type: object
          properties:
            assets:
              type: array
              items:
                $ref: "#/components/schemas/AssetSummary"
            line_items:
              type:
                - array
                - "null"
              items:
                $ref: "#/components/schemas/LineItemSummary"
            validation_findings:
              type:
                - array
                - "null"
              items:
                $ref: "#/components/schemas/ValidationFindingSummary"
            safe_error_code:
              type:
                - string
                - "null"
              description: >
                Non-sensitive failure classification code. Present only when
                the last processing attempt failed. Never contains receipt
                content or provider raw output.
            provenance_summary:
              type:
                - object
                - "null"
              description: >
                Safe extraction provenance (provider name, model ID, prompt
                version, schema version, attempt count). Never contains
                credentials, prompt text, or raw model output.
              properties:
                provider:
                  type: string
                model_id:
                  type: string
                prompt_version:
                  type: string
                schema_version:
                  type: string
                attempt_count:
                  type: integer

    # ── Human revision request schemas ────────────────────────────────────────
...
    CreateHumanRevisionRequest:
      type: object
      required:
        - expected_parent_revision_id
        - currency
        - total_minor
      properties:
        expected_parent_revision_id:
          $ref: "#/components/schemas/UUID"
          description: >
            Must equal the receipt's current_revision_id. The server rejects
            with 409 STALE_PARENT_REVISION if it does not match, preventing
            concurrent overwrites and ensuring exactly-once semantics.
        merchant_normalized:
          type:
            - string
            - "null"
          maxLength: 500
        purchase_datetime:
          type:
            - string
            - "null"
          format: date-time
        purchase_timezone:
          type:
            - string
            - "null"
          maxLength: 100
        currency:
          $ref: "#/components/schemas/Currency"
        subtotal_minor:
          $ref: "#/components/schemas/MinorUnits"
        tax_minor:
          $ref: "#/components/schemas/MinorUnits"
        tip_minor:
          $ref: "#/components/schemas/MinorUnits"
        discount_minor:
          $ref: "#/components/schemas/MinorUnits"
        total_minor:
          type: integer
          minimum: 0
          maximum: 9223372036854775807
          description: >
            Required receipt total in minor units. Must be arithmetically
            consistent with subtotal + tax + tip - discount within a 1-cent
            tolerance. Inconsistent totals return 422 without any write.
        line_items:
          type: array
          maxItems: 200
          items:
            $ref: "#/components/schemas/LineItemInput"
          description: >
            Complete replacement line-item snapshot. The server re-numbers
            items contiguously starting from 1; client ordinals are ignored.

    # ── Error schema ─────────────────────────────────────────────────────────
    ApiError:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          description: >
            Stable machine-readable error code. Never contains receipt
            content, credentials, or internal identifiers.
          example: "RECEIPT_NOT_FOUND"
        message:
          type: string
          description: Human-readable message safe for display. No private detail.
        request_id:
          type:
            - string
            - "null"
          description: Trace/correlation ID for support reference.
  responses:
    Unauthorized:
      description: >
        Missing or invalid Bearer token. Response contains no private detail.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error_code: "UNAUTHORIZED"
            message: "Authentication required."
    Forbidden:
      description: >
        Valid token for a non-allowlisted identity, or session version mismatch.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error_code: "FORBIDDEN"
            message: "Access denied."
    NotFound:
      description: Receipt or asset not found for this owner.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error_code: "RECEIPT_NOT_FOUND"
            message: "Receipt not found."
    UnprocessableEntity:
      description: Request schema validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Conflict:
      description: >
        State conflict (e.g., receipt already finalized, retry not permitted).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
paths:
...
  /api/v1/receipts:
    post:
      operationId: createReceipt
      summary: Create a receipt and obtain upload capabilities
      description: |
        Creates a new receipt and returns short-lived signed upload URLs for
        each declared asset (OBJ-02).

        **Idempotency (A-01):** The `client_submission_key` is a Version 4 UUID
        generated by the client using `crypto.randomUUID()` before this call.
        If a receipt with the same `client_submission_key` already exists for
        this owner, the existing receipt and new upload capabilities are returned
        with HTTP 200. Both concurrent callers receive a 2xx; only one receipt
        is created. Test: two concurrent POSTs → two 2xx, one receipt.

        **Key must be unique per owner.** Do not reuse a key for a different
        receipt; the server returns the original receipt without creating a new one.

        **Upload capabilities** are bearer secrets with narrow object scope and
        short expiry. Never log, cache in service-worker, or include in analytics.
      tags: [receipts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_submission_key
                - expected_asset_count
                - assets
              properties:
                client_submission_key:
                  $ref: "#/components/schemas/UUID"
                  description: >
                    Client-generated Version 4 UUID. Must be produced by
                    `crypto.randomUUID()` (Web Crypto API) in a secure context
                    before this request is sent. Never reuse across receipts.
                expected_asset_count:
                  type: integer
                  minimum: 1
                  maximum: 10
                  description: Total number of images in this receipt.
                financial_context:
                  $ref: "#/components/schemas/FinancialContext"
                  default: personal
                captured_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                  description: >
                    Client-reported capture time when reliably known.
                    Optional; stored as provided without invention of timezone.
                assets:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items:
                    $ref: "#/components/schemas/AssetInput"
                  description: >
                    Ordered asset declarations. Length must match
                    `expected_asset_count`. Ordinals must be contiguous
                    starting from 1.
      responses:
        "201":
          description: Receipt created. Upload capabilities returned.
          content:
            application/json:
              schema:
                type: object
                required:
                  - receipt_id
                  - processing_status
                  - upload_capabilities
                properties:
                  receipt_id:
                    $ref: "#/components/schemas/UUID"
                  processing_status:
                    $ref: "#/components/schemas/ProcessingStatus"
                  upload_capabilities:
                    type: array
                    items:
                      $ref: "#/components/schemas/UploadCapability"
        "200":
          description: >
            Receipt already exists for this `client_submission_key` and owner.
            Fresh upload capabilities are returned. Idempotent replay.
          content:
            application/json:
              schema:
                type: object
                required:
                  - receipt_id
                  - processing_status
                  - upload_capabilities
                properties:
                  receipt_id:
                    $ref: "#/components/schemas/UUID"
                  processing_status:
                    $ref: "#/components/schemas/ProcessingStatus"
                  upload_capabilities:
                    type: array
                    items:
                      $ref: "#/components/schemas/UploadCapability"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    get:
      operationId: listReceipts
      summary: List receipts for the authenticated owner
      description: |
        Returns receipts ordered newest first. Each row includes current
        processing/verification status and a current revision summary when available.

        This is not an analytics dashboard. Pagination uses an opaque cursor.
      tags: [receipts]
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type:
              - string
              - "null"
          description: Opaque pagination cursor from a previous response.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Maximum receipts to return.
      responses:
        "200":
          description: List of receipts.
          content:
            application/json:
              schema:
                type: object
                required:
                  - receipts
                properties:
                  receipts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReceiptListItem"
                  next_cursor:
                    type:
                      - string
                      - "null"
                    description: Cursor for the next page. Null when no more results.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/receipts/{receipt_id}:
    get:
      operationId: getReceipt
      summary: Get receipt detail including ordered evidence and current revision
      tags: [receipts]
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Full receipt detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/receipts/{receipt_id}/finalize:
    post:
      operationId: finalizeReceipt
      summary: Verify uploaded evidence and durably acknowledge the receipt
      description: |
        Verifies that all declared assets are present in private storage with
        matching object metadata (OBJ-03), persists durable receipt state, and
        enqueues an authenticated processing task (API-03, QUE-01).

        The response is sent only after evidence and receipt state are durable
        and the receipt is eligible for deterministic re-enqueue even if the
        first task dispatch fails. "Receipt saved" is never claimed without this.

        **Idempotency:** Safe to retry. A receipt already in `queued` or later
        state returns 200 without re-enqueuing. A receipt with evidence missing
        or invalid returns 422.

        **Verification steps (OBJ-03):** object existence, owner/receipt path
        match, ordered count, byte size within limits, allowed MIME type,
        decodable image content, `storage_generation` recorded.
      tags: [receipts]
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              description: No required body fields. Reserved for future finalization hints.
      responses:
        "200":
          description: Receipt acknowledged. Processing queued.
          content:
            application/json:
              schema:
                type: object
                required:
                  - receipt_id
                  - processing_status
                  - acknowledged_at
                properties:
                  receipt_id:
                    $ref: "#/components/schemas/UUID"
                  processing_status:
                    $ref: "#/components/schemas/ProcessingStatus"
                  acknowledged_at:
                    $ref: "#/components/schemas/Timestamp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: Evidence set is incomplete or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error_code: "EVIDENCE_INCOMPLETE"
                message: "One or more expected images are missing or could not be verified."
  /api/v1/receipts/{receipt_id}/retry-processing:
    post:
      operationId: retryProcessing
      summary: Re-enqueue a receipt for extraction after a retryable failure
      description: |
        Permitted only when the receipt's processing status is `retryable_failed`
        or `failed` (with supervisor approval). Creates a new processing attempt
        and re-enqueues. Idempotent: a receipt already in `queued` or `processing`
        returns 200 without creating a duplicate task.
      tags: [receipts]
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Receipt re-queued for processing.
          content:
            application/json:
              schema:
                type: object
                required:
                  - receipt_id
                  - processing_status
                properties:
                  receipt_id:
                    $ref: "#/components/schemas/UUID"
                  processing_status:
                    $ref: "#/components/schemas/ProcessingStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Retry not permitted for the current processing status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error_code: "RETRY_NOT_PERMITTED"
                message: "Receipt is not in a retryable state."
  /api/v1/receipts/{receipt_id}/assets/{asset_id}/download:
    post:
      operationId: getAssetDownloadCapability
      summary: Obtain a short-lived capability to download a receipt asset
      description: |
        Returns a short-lived server-authorized download URL for a specific
        receipt asset. The URL provides authenticated access to private object
        storage without making the bucket public (SEC-002, OBJ-01).

        The returned URL is a bearer secret — never log, include in analytics,
        or cache in a service worker. It must be used immediately for image
        display and discarded.
      tags: [receipts]
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: asset_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Short-lived download capability.
          content:
            application/json:
              schema:
                type: object
                required:
                  - download_url
                  - method
                  - expires_at
                properties:
                  download_url:
                    type: string
                    format: uri
                    description: >
                      Short-lived signed URL. Bearer secret — never log
                      or cache (OBJ-02, LOG-01).
                  method:
                    type: string
                    enum: ["GET"]
                  expires_at:
                    $ref: "#/components/schemas/Timestamp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/receipts/{receipt_id}/human-revisions:
    post:
      operationId: createHumanRevision
      summary: Submit a human correction and advance the receipt to human_verified
      description: |
        Creates an immutable `human` receipt revision whose parent is the receipt's
        current revision. On success the receipt's `current_revision_id` is
        atomically advanced and `verification_status` becomes `human_verified`.

        **Pre-conditions:**
        - Receipt exists and is owned by the caller.
        - `processing_status` is `extracted` and a current revision exists.
        - `expected_parent_revision_id` equals the receipt's `current_revision_id`.

        **Stale-write protection:** If two concurrent requests send the same
        `expected_parent_revision_id`, only the first succeeds. The second
        receives `409 STALE_PARENT_REVISION`; no duplicate revision is created.

        **Arithmetic enforcement:** The server validates that
        `subtotal + tax + tip - discount ≈ total` (1-minor-unit tolerance).
        Inconsistent totals return `422` without any write.

        **Immutability:** The original extractor revision, extraction run,
        raw model response, and evidence assets are never modified.

        **Lineage:** The new revision's `parent_revision_id` is set to the
        previous `current_revision_id`, forming an explicit acyclic chain.

        **Line-item re-numbering:** The server re-numbers submitted line items
        contiguously from 1. Client-supplied ordinals are ignored.

        **Validation findings:** Deterministic checks are re-run on the
        corrected snapshot and stored on the new revision.
      tags: [receipts]
      parameters:
        - name: receipt_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHumanRevisionRequest"
      responses:
        "200":
          description: >
            Human revision created. Receipt is now human_verified. Returns the
            updated receipt detail with source_type human and verification_status
            human_verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            Stale parent revision (STALE_PARENT_REVISION) or receipt not in a
            correctable state (INVALID_RECEIPT_STATE).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                stale:
                  value:
                    error_code: "STALE_PARENT_REVISION"
                    message: "The expected parent revision does not match the current revision."
                invalid_state:
                  value:
                    error_code: "INVALID_RECEIPT_STATE"
                    message: "Receipt is not in a state that allows human correction."
        "422":
          description: >
            Input validation failure. Invalid decimal strings, out-of-bounds
            values, bad currency code, or arithmetically inconsistent totals.
            No revision is created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error_code: "VALIDATION_ERROR"
                message: "Corrected totals or line-item arithmetic is inconsistent."

  # ── Internal worker routes ─────────────────────────────────────────────────
...